backend/requirements.txt
spacy
fastapi
uvicorn[standard]
pydantic
pymupdf
python-dotenv
sentence_transformers
numpy
langchain
langchain-community
python-multipart